requirements.txt
python-dotenv
Telethon~=1.27.0
cryptg>=0.4.0
tqdm>=4.64.1
